small refactor on disk autoincrement index, move some helper functions
